backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pdfminer.six==20231228
baml==0.4.4
python-multipart==0.0.9
pydantic==1.10.15
pytesseract==0.3.10
pdf2image==1.17.0
Pillow==10.3.0
